numen-modifications/pyproject.toml
[tool.poetry]
name = "numen-additions"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "numen_additions"}]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
